some comments

+++ core_0/src/text.h
@@ -1,6 +1,8 @@
 #include "font8x8_basic.h"
 #include <string.h>
 
+
+// Used to draw an arbitrary length string to any position at a specified memory at any color
 void draw_text_at_location(int x, int y, const char * text, int * memory_location, int color) {
 	unsigned int text_length = strlen(text);
 	char char_slice;
@@ -16,11 +18,13 @@ void draw_text_at_location(int x, int y, const char * text, int * memory_locatio
 	    for (unsigned int char_number = 0; char_number < text_length; char_number++) {
 	    	// the font8x8_basic uses the unicode values for array indexing so we can just pass in the
 	    	// actual char value
-	        char_slice = font8x8_basic[text[char_number]][row];
+	        char_slice = font8x8_basic[static_cast<int>(text[char_number])][row];
 
 	        for (int column = 0; column < 8; column++) {
 	        	// 1 if we should draw at this location, else 0
 	            pixel_val = !!(char_slice & (1 << (column)));
+
+	            // since pixel_val is binary, this ends up just being the colour or zero
 	            text_row[char_number * 8 + column] = pixel_val * color;
 	        }
 	    }
